fix(portal): move Header component
affects: @fremtind/portal
