+++ portal/src/components/index.ts
@@ -1,4 +1,4 @@
-import { Header } from "./Header/header-react/src";
+import { Header } from "./Header";
 import { Layout } from "./Layout/layout-react/src";
 import { ComponentLayout } from "./Layout/ComponentLayout";
 import { Menu } from "./Menu/Menu";
